pyproject.toml
[project]
name = "chess-analysis"
version = "0.1.0"
description = "Python dependancies to execute chess ingestion pipeline against GCP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.166.0",
    "google-cloud>=0.34.0",
    "google-cloud-bigquery>=3.31.0",
    "google-cloud-logging>=3.11.4",
    "google-cloud-storage>=3.1.0",
    "ipykernel>=6.29.5",
    "marimo>=0.11.30",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
]
